more package refactoring
